test/ginkgo-ext: check of skip flag to see if the test is focused
The current implementation of isTestFocused is only checking for the
FocusString. This causes some bugs in the test runs such as the
miss of some "AfterAll" executions since the ref count of all
"AfterEach"es is used by taking into account how many tests will be
executed. Checking how many tests will be executed is done by checking
if a particular test is "focused", when using a skip string, and the
test matches the test name, the test should not be marked as focused.

Signed-off-by: André Martins <andre@cilium.io>