gha: add clustermesh upgrade and downgrade tests
This commit introduces a new GitHub actions workflow which validates the
Cilium upgrade and downgrade paths when clustermesh is enabled, ensuring
that this process does not disrupt long living connections.

More specifically, the test initially deploys a mixed version cluster
mesh, with one cluster running the latest stable version of Cilium, and
the other the current tip of main. It performs a subset of connectivity
tests and deploys the application sensitive to connection disruption.
At this point, the first cluster is upgraded to the tip of main,
enabling kvstoremesh at the same time. Connectivity tests are again
executed, additionally checking that no long living connection was
dropped. Finally, the first cluster is downgraded again to the latest
stable version of Cilium, disabling kvstoremesh. Connectivity tests
and connection disruption checks are executed on more time.

To reduce the total amount of time required by this workflow, only a
limited subset of tests is enabled, while the full suite is run in
the conformance-clustermesh workflow.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>