ipcache: don't remove CIDR labels from prefixes with remote-node label
We would like to be able to select nodes by CIDR
selectors. So, we need to make sure the cidr labels are not forcibly
removed.

Now that remote nodes are allocated IDs from their own scope, the
datapath can still route packets correctly, and it is safe to add a CIDR
label and allocate an identity accordingly.

Signed-off-by: Casey Callendrello <cdc@isovalent.com>